feat(ssl): extract SSL into its own file
Extract the ssl options into their own file, and add it to both servers.
